refactor(cmd/internal): move package utils from cmd/internal to internal
