+++ internal/utils/extract.go
